round3/experiments/requirements.txt
opencv-python>=4.8
numpy>=1.24
matplotlib>=3.7
google-generativeai>=0.8
